Corregido error
El error lo que quitado convirtiéndolo a int y olvidándome del long. Hay que quitar el _stopped y hacer el invoke later en todas las cosas que cambien la interfaz, por ejemplo los mapas, pero los botones no. Por alguna razón que tengo que mirar, cuando ejecutamos solo va de uno en uno y debería de avanzar pero con los delays

+++ src/simulator/view/ControlPanel.java
@@ -124,7 +124,7 @@ public class ControlPanel extends JPanel implements TrafficSimObserver {
 				enableToolBar(true);
 				_thread = new Thread() { // lo creamos
 					public void run() {
-						run_sim((int) ticks.getValue(), (long) delay.getValue()); // TODO me esta dando error aqui
+						run_sim((int) ticks.getValue(), (int) delay.getValue()); // TODO me esta dando error aqui
 						_stopped = true;
 						enableToolBar(true);
 					}
@@ -187,7 +187,7 @@ public class ControlPanel extends JPanel implements TrafficSimObserver {
 		return button;
 	}
 	
-	private void run_sim(int n, long m) {
+	private void run_sim(int n, int m) {
 		while (n > 0 && !_stopped) {
 			try {
 				_ctrl.run(1);
